presentation-automator/backend/pyproject.toml
[build-system]
requires = ["setuptools>=42.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "presentation-automator-backend"
version = "0.1.0"
description = "Backend for the Presentation Automator system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "openai-agents>=0.0.1",
    "loguru>=0.7.0",
    "python-docx>=0.8.11",
    "PyPDF2>=3.0.0",
    "sse-starlette>=1.6.5",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "bcrypt>=4.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0"
] 